vo: sleep as much as possible before the vsync
The arbitrary value of 4ms at which we wake is just something that’s
big enough to account for variations of vsync times.

A much nicer solution would include automatically adjust this value so
that’s as late as possible for mpv to not miss the vsync while still
preserving good responsiveness (by sleeping as much as possible). Or
we could integrate APIs like DisplayLink on OS X so that we get waked
just before vsync (DisplayLink would automatically adjust the wake time
if it detects we arrived late for the last vsync).